Move pipeline support to dedicated package
