+++ src/main/java/jenkins/plugins/openstack/compute/JCloudsSlaveTemplate.java
@@ -212,7 +212,8 @@ public class JCloudsSlaveTemplate implements Describable<JCloudsSlaveTemplate>, 
         return provision(cloud, null);
     }
 
-    /*package*/ @Nonnull Server provision(@Nonnull JCloudsCloud cloud, @CheckForNull ServerScope scope) throws Openstack.ActionFailed {
+    @Restricted(NoExternalUse.class)
+    public @Nonnull Server provision(@Nonnull JCloudsCloud cloud, @CheckForNull ServerScope scope) throws Openstack.ActionFailed {
         final String serverName = name + "-" + new Random().nextInt(10000);
         final SlaveOptions opts = getEffectiveSlaveOptions();
         final ServerCreateBuilder builder = Builders.server();
